KWON 물품 등록

+++ src/main/java/com/tenco/toyproject/utils/DecimalUtil.java
@@ -4,7 +4,7 @@ import java.text.DecimalFormat;
 
 public class DecimalUtil {
 
-    public static String decimalToString(int amount) {
+    public static String decimalToString(Long amount) {
         DecimalFormat df = new DecimalFormat("#,###");
         String formatNumber = df.format(amount);
         return formatNumber;
